Funções/Usuarios/21/11

+++ database/migrations/2023_11_07_00006_create_telefone_usuarios_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('telefone_usuarios', function (Blueprint $table) {
-            $table->Integer('telefone')->nullable(false);
+            $table->bigInteger('telefone')->nullable(false);
 
             //Chave estrangeira de usuarios
             $table->unsignedBigInteger('Usuarios_idUsuario');
